Added planes
Since holes will not inherently be plated through holes, I tried to maintain connection to components entirely on one side. A few traces are on the top side, with the intention of them be soldered from above in addition to below. I also placed a few large vias, with the idea being the solder will join them together. Can't wait to get my CNC machine and try this out!